% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeFloodMetrics.R
\name{analyzeFloodMetrics}
\alias{analyzeFloodMetrics}
\title{Analyze Flood Metrics}
\usage{
analyzeFloodMetrics(FldFl, i, Min_AP, percentile)
}
\arguments{
\item{FldFl}{A dataframe containing the flood flow data.}

\item{i}{An identifier or index.}

\item{Min_AP}{Minimum Annual Precipitation for Annual Maximum Series.}

\item{percentile}{The ratio used in the percentile calculation.}
}
\value{
A dataframe with analyzed flood metrics.
}
\description{
This function analyzes flood metrics based on the provided flood flow data.
}
\examples{
\dontrun{
# Assuming you have a dataframe 'FldFl' with flood flow data, an identifier 'i', Min_AP, and percentile:
result_df <- analyzeFloodMetrics(FldFl, i, Min_AP, percentile)
}

}
