% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariateCopulaWithBestFit.R
\name{bivariateCopulaWithBestFit}
\alias{bivariateCopulaWithBestFit}
\title{Bivariate Copula Analysis with Best Fit Distributions}
\usage{
bivariateCopulaWithBestFit(X1, X2)
}
\arguments{
\item{X1}{A numeric vector representing the first data set.}

\item{X2}{A numeric vector representing the second data set.}
}
\value{
A list containing the results of the copula analysis.
}
\description{
This function fits two datasets to multiple distributions, selects the best distribution based on AIC,
and then performs a bivariate copula analysis.
}
\examples{
\dontrun{
# Assuming you have two vectors 'X1' and 'X2' and a function 'fitMultipleDistributions':
result <- bivariateCopulaWithBestFit(X1, X2, fitMultipleDistributions)
}


}
