% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDistributions.R
\name{compareDistributions}
\alias{compareDistributions}
\title{Compare Best-Fit and GEV Distributions}
\usage{
compareDistributions(
  sample1_properties,
  sample2_properties,
  p = c(0.05, seq(0.1, 0.9, 0.005), 0.95),
  output_filename = "CDF_comparison.svg"
)
}
\arguments{
\item{sample1_properties}{A named list containing the properties of the first sample (distribution and its parameters).}

\item{sample2_properties}{A named list containing the properties of the second sample (distribution and its parameters).}

\item{p}{A numeric vector of percentiles.}

\item{output_filename}{The filename for the SVG output plot.}
}
\value{
A dataframe containing the KS test statistic and p-value.
}
\description{
This function compares the best-fit and GEV distributions using the KS test and plots their CDFs.
}
