% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulativeDensityFunction.R
\name{cumulativeDensityFunction}
\alias{cumulativeDensityFunction}
\title{Cumulative Density of Specified Distribution}
\usage{
cumulativeDensityFunction(dbn, P, shp, sc)
}
\arguments{
\item{dbn}{A character string specifying the distribution. Valid options are "norm", "lnorm", "gamma", "weibull", and "llogis".}

\item{P}{A numeric value or vector for which the cumulative density is to be computed.}

\item{shp}{Shape parameter for the distribution.}

\item{sc}{Scale parameter for the distribution.}
}
\value{
A numeric value or vector representing the cumulative density for the specified distribution and parameters.
}
\description{
This function computes the cumulative density value for a given distribution and parameters.
}
\examples{
\dontrun{
cumulativeDensityFunction("norm", 0, 0, 1)
cumulativeDensityFunction("gamma", 1, 2, 1)
}

}
