% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFitStatistics.R
\name{extractFitStatistics}
\alias{extractFitStatistics}
\title{Fit Statistics Extraction}
\usage{
extractFitStatistics(obj)
}
\arguments{
\item{obj}{Object that is obtained after fitting a set of data to fitdist function offitdistplus library}
}
\value{
Shape, Scale, SDShape, SDscale, Loglikelihood and AIC of the distribution
}
\description{
Function used to obtain coefficients and Metrics for each distribution

Takes the object fitted using fitdistrplus

Extracts - Shape, Scale, SDShape, SDscale, Loglikelihood and AIC
}
\examples{
\dontrun{
# Generate some sample data
data <- rnorm(100)

# Fit the data using fitdist from the fitdistrplus package
fit <- fitdistrplus::fitdist(data, "norm")

# Extract fit statistics
extractFitStatistics(fit)
}
#'
}
