% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPeakFlow.R
\name{extractPeakFlow}
\alias{extractPeakFlow}
\title{Extract Peak Flow and Time to Peak}
\usage{
extractPeakFlow(df)
}
\arguments{
\item{df}{A dataframe containing the streamflow data with columns 'Floodidx', 'FloodQ', and 'Date'.}
}
\value{
A dataframe with peak flows and their corresponding dates.
}
\description{
This function extracts the peak flow and the corresponding time to peak from a given dataframe.
}
\examples{
\dontrun{
# Assuming you have a dataframe 'df' with streamflow data:
result_df <- extractPeakFlow(df)
}

}
