% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAndTestDistributions.R
\name{fitAndTestDistributions}
\alias{fitAndTestDistributions}
\title{Fit Multiple Distributions and Test Goodness of Fit}
\usage{
fitAndTestDistributions(Stn, StnID)
}
\arguments{
\item{Stn}{A dataframe containing the peak flow data in a column named 'peak_va'.}

\item{StnID}{A numeric or character value representing the station ID.}
}
\value{
A dataframe containing fit parameters and test statistics for each distribution.
}
\description{
This function fits multiple distributions to peak flow data, computes empirical probabilities, and tests the goodness of fit using the KS and AD tests.
}
\examples{
\dontrun{
# Assuming you have a dataframe 'Stn' and a station ID '08041500':
fitAndTestDistributions(Stn, '08041500')
}

}
