% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitExpAndGammaDistributions.R
\name{fitExpAndGammaDistributions}
\alias{fitExpAndGammaDistributions}
\title{Fit Exponential and Gamma Distributions to Data}
\usage{
fitExpAndGammaDistributions(X)
}
\arguments{
\item{X}{A numeric vector of data to be fitted into the distributions.}
}
\value{
A data frame containing the fit statistics for each distribution, including their shape, scale,
standard errors for shape and scale, Log-likelihood, and AIC value.
}
\description{
This function fits both exponential and gamma distributions to the provided data using Maximum Likelihood Estimation (MLE).
}
\examples{
\dontrun{
fitExpAndGammaDistributions(data)
}

}
