% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitExtremeDistributions.R
\name{fitExtremeDistributions}
\alias{fitExtremeDistributions}
\title{Fit Extreme Distributions (GEV and Gumbel) to Peak Flow Data}
\usage{
fitExtremeDistributions(Stn)
}
\arguments{
\item{Stn}{A dataframe containing the peak flow data in a column named 'peak_va'.}
}
\value{
A dataframe with fitted parameters for the GEV and Gumbel distributions.
}
\description{
This function fits the Generalized Extreme Value (GEV) and Gumbel distributions to peak flow data from a given station.
}
\examples{
\dontrun{
# Assuming you have a dataframe 'Stn' with peak flow data:
results <- fitExtremeDistributions(Stn)
}

}
