% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMultipleDistributions.R
\name{fitMultipleDistributions}
\alias{fitMultipleDistributions}
\title{Fit Multiple Distributions to Data}
\usage{
fitMultipleDistributions(X)
}
\arguments{
\item{X}{A numeric vector of data to be fitted into various distributions.}
}
\value{
A data frame containing the fit statistics for each distribution, including their shape, scale,
Log-likelihood, and AIC value.
}
\description{
This function fits five candidate distributions to the provided data using Maximum Likelihood Estimation (MLE).
The distributions fitted are: normal, lognormal, weibull, gamma, and logistic.
}
\examples{
\dontrun{
fitMultipleDistributions(data)
}

}
