% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floodCharacteristics.R
\name{floodCharacteristics}
\alias{floodCharacteristics}
\title{Compute Flood Characteristics}
\usage{
floodCharacteristics(df, StnID, volume_threshold = 4355990)
}
\arguments{
\item{df}{Dataframe with flood data.}

\item{StnID}{ID of the Station}

\item{volume_threshold}{The volume threshold to filter out floods (default is set to 4355990).}
}
\value{
A dataframe with flood characteristics.
}
\description{
This function computes flood characteristics from a given flood data dataframe.
}
\examples{
\dontrun{
# Assuming you have a dataframe 'df' with flood data:
result_df <- floodCharacteristics(df, '08041500')
}

}
