% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowDataDiagnostics.R
\name{flowDataDiagnostics}
\alias{flowDataDiagnostics}
\title{Analyze Flow Data for Trends, Change Points, and Stationarity}
\usage{
flowDataDiagnostics(pf)
}
\arguments{
\item{pf}{A numeric vector or time series representing flow data.}
}
\value{
A list containing the results of each of the tests performed.
}
\description{
This function conducts a series of tests on flow data to detect trends, change points, and assess stationarity. The tests performed include the Mann-Kendall Test, Pettitt Test, Augmented Dickey-Fuller Test, KPSS Test, and the Phillips-Perron Unit Root Test.
}
\examples{
\dontrun{
  data <- rnorm(100)  # Example data
  results <- flowDataDiagnostics(data)
  print(results)
}

}
\references{
McLeod, A. I. (2005). Kendall rank correlation and Mann-Kendall trend test. R Package Kendall, 602, 1-10.
}
