% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenParPois.R
\name{GenParPois}
\alias{GenParPois}
\title{Generalized Pareto Distribution for Flood Metrics}
\usage{
GenParPois(i, FldFl, Min_AP, thres)
}
\arguments{
\item{i}{An identifier or index.}

\item{FldFl}{A dataframe containing the flood flow data.}

\item{Min_AP}{Minimum Annual Precipitation for Annual Maximum Series.}

\item{thres}{The threshold for the Generalized Pareto Distribution.}
}
\value{
A dataframe with parameters and statistics related to the fitted Generalized Pareto Distribution.
}
\description{
This function fits a Generalized Pareto Distribution to flood flow data.
}
\examples{
\dontrun{
# Assuming you have a dataframe 'FldFl' with flood flow data, an identifier 'i', Min_AP, and threshold 'thres':
result_df <- GenParPois(i, FldFl, Min_AP, thres)
}

}
