% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEVPars.R
\name{GEVPars}
\alias{GEVPars}
\title{Compute GEV Parameters}
\usage{
GEVPars(a)
}
\arguments{
\item{a}{A dataframe containing the columns 'GP_Location', 'GP_scale', 'GP_shape', and 'Pois_Lam'.}
}
\value{
A dataframe with added columns for 'GEV_loc', 'GEV_scale', and 'GEV_shp'.
}
\description{
This function calculates the GEV parameters based on the provided dataframe.
}
\examples{
\dontrun{
# Assuming you have a dataframe 'FldFl' with flood flow data, an identifier 'i', Min_AP, and percentile:
result_df <- GEVPars(df)
}

}
