% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gringorten2DBivariateEmpirical.R
\name{gringorten2DBivariateEmpirical}
\alias{gringorten2DBivariateEmpirical}
\title{Bivariate Empirical Distribution Using Gringorten's Formula}
\usage{
gringorten2DBivariateEmpirical(idx, X)
}
\arguments{
\item{idx}{An integer index indicating which row of the data matrix \code{X} should be used as the reference point.}

\item{X}{A numeric matrix or data frame where each row represents a bivariate data point.}
}
\value{
A numeric value representing the bivariate empirical distribution value for the specified data point.
}
\description{
This function computes the bivariate empirical distribution value for a given data point using Gringorten's plotting formula.
}
\examples{
\dontrun{
data <- matrix(rnorm(100), ncol = 2)
gringorten2DBivariateEmpirical(10, data)
}

}
