% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independenceCriteriaCheck.R
\name{independenceCriteriaCheck}
\alias{independenceCriteriaCheck}
\title{Independence Criteria Check on Flood}
\usage{
independenceCriteriaCheck(QT, Ar, daily_flow, flow_threshold = 10)
}
\arguments{
\item{QT}{A dataframe with peak flows and corresponding time obtained after baseflow separation.}

\item{Ar}{Area of the watershed in square kilometers.}

\item{daily_flow}{Dataframe with daily flow data.}

\item{flow_threshold}{The value to filter out flows (default is set to 10).}
}
\value{
A dataframe with peak flows and their corresponding dates after applying the independence criteria checks.
}
\description{
This function applies independence criteria checks on flood data and filters out flows below a given threshold.
}
\examples{
\dontrun{
#Assuming you have dataframes 'QT' with peak flows, 'daily_flow' with daily flow data, and area 'Ar':
result_df <- independenceCriteriaCheck(QT, Ar, daily_flow)
}

}
