% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAnnualPeakFlow.R
\name{plotAnnualPeakFlow}
\alias{plotAnnualPeakFlow}
\title{Time Series Plot of Annual Peak Flow}
\usage{
plotAnnualPeakFlow(df, st_yr, end_yr)
}
\arguments{
\item{df}{A dataframe with annual peak flow in a column labeled peak_va.}

\item{st_yr}{Start Year.}

\item{end_yr}{End Year.}
}
\value{
Time Series Plot
}
\description{
This function takes annual peak flow data and plots it as a time series for a specified range of years.
}
\examples{
\dontrun{
plotAnnualPeakFlow(df,1980,2022)
}

}
