% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAutocorrelations.R
\name{plotAutocorrelations}
\alias{plotAutocorrelations}
\title{Plot Autocorrelation and Partial Autocorrelation}
\usage{
plotAutocorrelations(df)
}
\arguments{
\item{df}{A numeric vector or time series.}
}
\value{
Autocorrelation Function (ACF) and Partial Autocorrelation Function (PACF) plots.
}
\description{
This function generates the Autocorrelation Function (ACF) and 'Partial Autocorrelation Function (PACF) plots for the given data.
}
\examples{
\dontrun{
plotAutocorrelations(data)
}

}
